course teacher_id nullable

+++ database/migrations/2022_08_05_180648_add_nullable_to_teacher_to_courses_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('courses', function (Blueprint $table) {
-            $table->unsignedInteger('teacher_id')->nullable()->change();
+            $table->unsignedBigInteger('teacher_id')->nullable()->change();
 
         });
     }
